Split agent and gather tests into multiple files
