+++ agent_candidate_pair_test.go
@@ -13,6 +13,24 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestNoBestAvailableCandidatePairAfterAgentConstruction(t *testing.T) {
+	agent := setupTest(t)
+
+	require.Nil(t, agent.getBestAvailableCandidatePair())
+
+	tearDownTest(t, agent)
+}
+
+func setupTest(t *testing.T) *Agent {
+	agent, err := NewAgent(&AgentConfig{})
+	require.NoError(t, err)
+	return agent
+}
+
+func tearDownTest(t *testing.T, agent *Agent) {
+	require.NoError(t, agent.Close())
+}
+
 func TestAgentGetBestValidCandidatePair(t *testing.T) {
 	f := setupTestAgentGetBestValidCandidatePair(t)
