+++ agent_stats_test.go
@@ -1,0 +1,349 @@
+// SPDX-FileCopyrightText: 2023 The Pion community <https://pion.ly>
+// SPDX-License-Identifier: MIT
+
+//go:build !js
+// +build !js
+
+package ice
+
+import (
+	"testing"
+	"time"
+
+	"github.com/pion/transport/v2/test"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCandidatePairStats(t *testing.T) {
+	report := test.CheckRoutines(t)
+	defer report()
+
+	// Avoid deadlocks?
+	defer test.TimeOut(1 * time.Second).Stop()
+
+	a, err := NewAgent(&AgentConfig{})
+	if err != nil {
+		t.Fatalf("Failed to create agent: %s", err)
+	}
+
+	hostConfig := &CandidateHostConfig{
+		Network:   "udp",
+		Address:   "192.168.1.1",
+		Port:      19216,
+		Component: 1,
+	}
+	hostLocal, err := NewCandidateHost(hostConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct local host candidate: %s", err)
+	}
+
+	relayConfig := &CandidateRelayConfig{
+		Network:   "udp",
+		Address:   "1.2.3.4",
+		Port:      2340,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43210,
+	}
+	relayRemote, err := NewCandidateRelay(relayConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote relay candidate: %s", err)
+	}
+
+	srflxConfig := &CandidateServerReflexiveConfig{
+		Network:   "udp",
+		Address:   "10.10.10.2",
+		Port:      19218,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43212,
+	}
+	srflxRemote, err := NewCandidateServerReflexive(srflxConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote srflx candidate: %s", err)
+	}
+
+	prflxConfig := &CandidatePeerReflexiveConfig{
+		Network:   "udp",
+		Address:   "10.10.10.2",
+		Port:      19217,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43211,
+	}
+	prflxRemote, err := NewCandidatePeerReflexive(prflxConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote prflx candidate: %s", err)
+	}
+
+	hostConfig = &CandidateHostConfig{
+		Network:   "udp",
+		Address:   "1.2.3.5",
+		Port:      12350,
+		Component: 1,
+	}
+	hostRemote, err := NewCandidateHost(hostConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote host candidate: %s", err)
+	}
+
+	for _, remote := range []Candidate{relayRemote, srflxRemote, prflxRemote, hostRemote} {
+		p := a.findPair(hostLocal, remote)
+
+		if p == nil {
+			a.addPair(hostLocal, remote)
+		}
+	}
+
+	p := a.findPair(hostLocal, prflxRemote)
+	p.state = CandidatePairStateFailed
+
+	stats := a.GetCandidatePairsStats()
+	if len(stats) != 4 {
+		t.Fatal("expected 4 candidate pairs stats")
+	}
+
+	var relayPairStat, srflxPairStat, prflxPairStat, hostPairStat CandidatePairStats
+
+	for _, cps := range stats {
+		if cps.LocalCandidateID != hostLocal.ID() {
+			t.Fatal("invalid local candidate id")
+		}
+		switch cps.RemoteCandidateID {
+		case relayRemote.ID():
+			relayPairStat = cps
+		case srflxRemote.ID():
+			srflxPairStat = cps
+		case prflxRemote.ID():
+			prflxPairStat = cps
+		case hostRemote.ID():
+			hostPairStat = cps
+		default:
+			t.Fatal("invalid remote candidate ID")
+		}
+	}
+
+	if relayPairStat.RemoteCandidateID != relayRemote.ID() {
+		t.Fatal("missing host-relay pair stat")
+	}
+
+	if srflxPairStat.RemoteCandidateID != srflxRemote.ID() {
+		t.Fatal("missing host-srflx pair stat")
+	}
+
+	if prflxPairStat.RemoteCandidateID != prflxRemote.ID() {
+		t.Fatal("missing host-prflx pair stat")
+	}
+
+	if hostPairStat.RemoteCandidateID != hostRemote.ID() {
+		t.Fatal("missing host-host pair stat")
+	}
+
+	if prflxPairStat.State != CandidatePairStateFailed {
+		t.Fatalf("expected host-prflx pair to have state failed, it has state %s instead",
+			prflxPairStat.State.String())
+	}
+
+	assert.NoError(t, a.Close())
+}
+
+func TestLocalCandidateStats(t *testing.T) {
+	report := test.CheckRoutines(t)
+	defer report()
+
+	// Avoid deadlocks?
+	defer test.TimeOut(1 * time.Second).Stop()
+
+	a, err := NewAgent(&AgentConfig{})
+	if err != nil {
+		t.Fatalf("Failed to create agent: %s", err)
+	}
+
+	hostConfig := &CandidateHostConfig{
+		Network:   "udp",
+		Address:   "192.168.1.1",
+		Port:      19216,
+		Component: 1,
+	}
+	hostLocal, err := NewCandidateHost(hostConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct local host candidate: %s", err)
+	}
+
+	srflxConfig := &CandidateServerReflexiveConfig{
+		Network:   "udp",
+		Address:   "192.168.1.1",
+		Port:      19217,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43212,
+	}
+	srflxLocal, err := NewCandidateServerReflexive(srflxConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct local srflx candidate: %s", err)
+	}
+
+	a.localCandidates[NetworkTypeUDP4] = []Candidate{hostLocal, srflxLocal}
+
+	localStats := a.GetLocalCandidatesStats()
+	if len(localStats) != 2 {
+		t.Fatalf("expected 2 local candidates stats, got %d instead", len(localStats))
+	}
+
+	var hostLocalStat, srflxLocalStat CandidateStats
+	for _, stats := range localStats {
+		var candidate Candidate
+		switch stats.ID {
+		case hostLocal.ID():
+			hostLocalStat = stats
+			candidate = hostLocal
+		case srflxLocal.ID():
+			srflxLocalStat = stats
+			candidate = srflxLocal
+		default:
+			t.Fatal("invalid local candidate ID")
+		}
+
+		if stats.CandidateType != candidate.Type() {
+			t.Fatal("invalid stats CandidateType")
+		}
+
+		if stats.Priority != candidate.Priority() {
+			t.Fatal("invalid stats CandidateType")
+		}
+
+		if stats.IP != candidate.Address() {
+			t.Fatal("invalid stats IP")
+		}
+	}
+
+	if hostLocalStat.ID != hostLocal.ID() {
+		t.Fatal("missing host local stat")
+	}
+
+	if srflxLocalStat.ID != srflxLocal.ID() {
+		t.Fatal("missing srflx local stat")
+	}
+
+	assert.NoError(t, a.Close())
+}
+
+func TestRemoteCandidateStats(t *testing.T) {
+	report := test.CheckRoutines(t)
+	defer report()
+
+	// Avoid deadlocks?
+	defer test.TimeOut(1 * time.Second).Stop()
+
+	a, err := NewAgent(&AgentConfig{})
+	if err != nil {
+		t.Fatalf("Failed to create agent: %s", err)
+	}
+
+	relayConfig := &CandidateRelayConfig{
+		Network:   "udp",
+		Address:   "1.2.3.4",
+		Port:      12340,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43210,
+	}
+	relayRemote, err := NewCandidateRelay(relayConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote relay candidate: %s", err)
+	}
+
+	srflxConfig := &CandidateServerReflexiveConfig{
+		Network:   "udp",
+		Address:   "10.10.10.2",
+		Port:      19218,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43212,
+	}
+	srflxRemote, err := NewCandidateServerReflexive(srflxConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote srflx candidate: %s", err)
+	}
+
+	prflxConfig := &CandidatePeerReflexiveConfig{
+		Network:   "udp",
+		Address:   "10.10.10.2",
+		Port:      19217,
+		Component: 1,
+		RelAddr:   "4.3.2.1",
+		RelPort:   43211,
+	}
+	prflxRemote, err := NewCandidatePeerReflexive(prflxConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote prflx candidate: %s", err)
+	}
+
+	hostConfig := &CandidateHostConfig{
+		Network:   "udp",
+		Address:   "1.2.3.5",
+		Port:      12350,
+		Component: 1,
+	}
+	hostRemote, err := NewCandidateHost(hostConfig)
+	if err != nil {
+		t.Fatalf("Failed to construct remote host candidate: %s", err)
+	}
+
+	a.remoteCandidates[NetworkTypeUDP4] = []Candidate{relayRemote, srflxRemote, prflxRemote, hostRemote}
+
+	remoteStats := a.GetRemoteCandidatesStats()
+	if len(remoteStats) != 4 {
+		t.Fatalf("expected 4 remote candidates stats, got %d instead", len(remoteStats))
+	}
+	var relayRemoteStat, srflxRemoteStat, prflxRemoteStat, hostRemoteStat CandidateStats
+	for _, stats := range remoteStats {
+		var candidate Candidate
+		switch stats.ID {
+		case relayRemote.ID():
+			relayRemoteStat = stats
+			candidate = relayRemote
+		case srflxRemote.ID():
+			srflxRemoteStat = stats
+			candidate = srflxRemote
+		case prflxRemote.ID():
+			prflxRemoteStat = stats
+			candidate = prflxRemote
+		case hostRemote.ID():
+			hostRemoteStat = stats
+			candidate = hostRemote
+		default:
+			t.Fatal("invalid remote candidate ID")
+		}
+
+		if stats.CandidateType != candidate.Type() {
+			t.Fatal("invalid stats CandidateType")
+		}
+
+		if stats.Priority != candidate.Priority() {
+			t.Fatal("invalid stats CandidateType")
+		}
+
+		if stats.IP != candidate.Address() {
+			t.Fatal("invalid stats IP")
+		}
+	}
+
+	if relayRemoteStat.ID != relayRemote.ID() {
+		t.Fatal("missing relay remote stat")
+	}
+
+	if srflxRemoteStat.ID != srflxRemote.ID() {
+		t.Fatal("missing srflx remote stat")
+	}
+
+	if prflxRemoteStat.ID != prflxRemote.ID() {
+		t.Fatal("missing prflx remote stat")
+	}
+
+	if hostRemoteStat.ID != hostRemote.ID() {
+		t.Fatal("missing host remote stat")
+	}
+
+	assert.NoError(t, a.Close())
+}
