+++ agent_test.go
@@ -9,7 +9,6 @@ package ice
 import (
 	"context"
 	"errors"
-	"fmt"
 	"net"
 	"strconv"
 	"sync"
@@ -294,67 +293,6 @@ func TestConnectivityOnStartup(t *testing.T) {
 	}
 }
 
-func TestConnectivityLite(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 30)
-	defer lim.Stop()
-
-	stunServerURL := &stun.URI{
-		Scheme: SchemeTypeSTUN,
-		Host:   "1.2.3.4",
-		Port:   3478,
-		Proto:  stun.ProtoTypeUDP,
-	}
-
-	natType := &vnet.NATType{
-		MappingBehavior:   vnet.EndpointIndependent,
-		FilteringBehavior: vnet.EndpointIndependent,
-	}
-	v, err := buildVNet(natType, natType)
-	require.NoError(t, err, "should succeed")
-	defer v.close()
-
-	aNotifier, aConnected := onConnected()
-	bNotifier, bConnected := onConnected()
-
-	cfg0 := &AgentConfig{
-		Urls:             []*stun.URI{stunServerURL},
-		NetworkTypes:     supportedNetworkTypes(),
-		MulticastDNSMode: MulticastDNSModeDisabled,
-		Net:              v.net0,
-	}
-
-	aAgent, err := NewAgent(cfg0)
-	require.NoError(t, err)
-	require.NoError(t, aAgent.OnConnectionStateChange(aNotifier))
-
-	cfg1 := &AgentConfig{
-		Urls:             []*stun.URI{},
-		Lite:             true,
-		CandidateTypes:   []CandidateType{CandidateTypeHost},
-		NetworkTypes:     supportedNetworkTypes(),
-		MulticastDNSMode: MulticastDNSModeDisabled,
-		Net:              v.net1,
-	}
-
-	bAgent, err := NewAgent(cfg1)
-	require.NoError(t, err)
-	require.NoError(t, bAgent.OnConnectionStateChange(bNotifier))
-
-	aConn, bConn := connectWithVNet(aAgent, bAgent)
-
-	// Ensure pair selected
-	// Note: this assumes ConnectionStateConnected is thrown after selecting the final pair
-	<-aConnected
-	<-bConnected
-
-	if !closePipe(t, aConn, bConn) {
-		return
-	}
-}
-
 func TestInboundValidity(t *testing.T) {
 	report := test.CheckRoutines(t)
 	defer report()
@@ -535,73 +473,6 @@ func TestInvalidAgentStarts(t *testing.T) {
 	assert.NoError(t, a.Close())
 }
 
-// Assert that Agent emits Connecting/Connected/Disconnected/Failed/Closed messages
-func TestConnectionStateCallback(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 5)
-	defer lim.Stop()
-
-	disconnectedDuration := time.Second
-	failedDuration := time.Second
-	KeepaliveInterval := time.Duration(0)
-
-	cfg := &AgentConfig{
-		Urls:                []*stun.URI{},
-		NetworkTypes:        supportedNetworkTypes(),
-		DisconnectedTimeout: &disconnectedDuration,
-		FailedTimeout:       &failedDuration,
-		KeepaliveInterval:   &KeepaliveInterval,
-	}
-
-	aAgent, err := NewAgent(cfg)
-	if err != nil {
-		t.Error(err)
-	}
-
-	bAgent, err := NewAgent(cfg)
-	if err != nil {
-		t.Error(err)
-	}
-
-	isChecking := make(chan interface{})
-	isConnected := make(chan interface{})
-	isDisconnected := make(chan interface{})
-	isFailed := make(chan interface{})
-	isClosed := make(chan interface{})
-	err = aAgent.OnConnectionStateChange(func(c ConnectionState) {
-		switch c {
-		case ConnectionStateChecking:
-			close(isChecking)
-		case ConnectionStateConnected:
-			close(isConnected)
-		case ConnectionStateDisconnected:
-			close(isDisconnected)
-		case ConnectionStateFailed:
-			close(isFailed)
-		case ConnectionStateClosed:
-			close(isClosed)
-		default:
-		}
-	})
-	if err != nil {
-		t.Error(err)
-	}
-
-	connect(aAgent, bAgent)
-
-	<-isChecking
-	<-isConnected
-	<-isDisconnected
-	<-isFailed
-
-	assert.NoError(t, aAgent.Close())
-	assert.NoError(t, bAgent.Close())
-
-	<-isClosed
-}
-
 func TestInvalidGather(t *testing.T) {
 	t.Run("Gather with no OnCandidate should error", func(t *testing.T) {
 		a, err := NewAgent(&AgentConfig{})
@@ -615,340 +486,6 @@ func TestInvalidGather(t *testing.T) {
 		}
 		assert.NoError(t, a.Close())
 	})
-}
-
-func TestCandidatePairStats(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	// Avoid deadlocks?
-	defer test.TimeOut(1 * time.Second).Stop()
-
-	a, err := NewAgent(&AgentConfig{})
-	if err != nil {
-		t.Fatalf("Failed to create agent: %s", err)
-	}
-
-	hostConfig := &CandidateHostConfig{
-		Network:   "udp",
-		Address:   "192.168.1.1",
-		Port:      19216,
-		Component: 1,
-	}
-	hostLocal, err := NewCandidateHost(hostConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct local host candidate: %s", err)
-	}
-
-	relayConfig := &CandidateRelayConfig{
-		Network:   "udp",
-		Address:   "1.2.3.4",
-		Port:      2340,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43210,
-	}
-	relayRemote, err := NewCandidateRelay(relayConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote relay candidate: %s", err)
-	}
-
-	srflxConfig := &CandidateServerReflexiveConfig{
-		Network:   "udp",
-		Address:   "10.10.10.2",
-		Port:      19218,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43212,
-	}
-	srflxRemote, err := NewCandidateServerReflexive(srflxConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote srflx candidate: %s", err)
-	}
-
-	prflxConfig := &CandidatePeerReflexiveConfig{
-		Network:   "udp",
-		Address:   "10.10.10.2",
-		Port:      19217,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43211,
-	}
-	prflxRemote, err := NewCandidatePeerReflexive(prflxConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote prflx candidate: %s", err)
-	}
-
-	hostConfig = &CandidateHostConfig{
-		Network:   "udp",
-		Address:   "1.2.3.5",
-		Port:      12350,
-		Component: 1,
-	}
-	hostRemote, err := NewCandidateHost(hostConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote host candidate: %s", err)
-	}
-
-	for _, remote := range []Candidate{relayRemote, srflxRemote, prflxRemote, hostRemote} {
-		p := a.findPair(hostLocal, remote)
-
-		if p == nil {
-			a.addPair(hostLocal, remote)
-		}
-	}
-
-	p := a.findPair(hostLocal, prflxRemote)
-	p.state = CandidatePairStateFailed
-
-	stats := a.GetCandidatePairsStats()
-	if len(stats) != 4 {
-		t.Fatal("expected 4 candidate pairs stats")
-	}
-
-	var relayPairStat, srflxPairStat, prflxPairStat, hostPairStat CandidatePairStats
-
-	for _, cps := range stats {
-		if cps.LocalCandidateID != hostLocal.ID() {
-			t.Fatal("invalid local candidate id")
-		}
-		switch cps.RemoteCandidateID {
-		case relayRemote.ID():
-			relayPairStat = cps
-		case srflxRemote.ID():
-			srflxPairStat = cps
-		case prflxRemote.ID():
-			prflxPairStat = cps
-		case hostRemote.ID():
-			hostPairStat = cps
-		default:
-			t.Fatal("invalid remote candidate ID")
-		}
-	}
-
-	if relayPairStat.RemoteCandidateID != relayRemote.ID() {
-		t.Fatal("missing host-relay pair stat")
-	}
-
-	if srflxPairStat.RemoteCandidateID != srflxRemote.ID() {
-		t.Fatal("missing host-srflx pair stat")
-	}
-
-	if prflxPairStat.RemoteCandidateID != prflxRemote.ID() {
-		t.Fatal("missing host-prflx pair stat")
-	}
-
-	if hostPairStat.RemoteCandidateID != hostRemote.ID() {
-		t.Fatal("missing host-host pair stat")
-	}
-
-	if prflxPairStat.State != CandidatePairStateFailed {
-		t.Fatalf("expected host-prflx pair to have state failed, it has state %s instead",
-			prflxPairStat.State.String())
-	}
-
-	assert.NoError(t, a.Close())
-}
-
-func TestLocalCandidateStats(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	// Avoid deadlocks?
-	defer test.TimeOut(1 * time.Second).Stop()
-
-	a, err := NewAgent(&AgentConfig{})
-	if err != nil {
-		t.Fatalf("Failed to create agent: %s", err)
-	}
-
-	hostConfig := &CandidateHostConfig{
-		Network:   "udp",
-		Address:   "192.168.1.1",
-		Port:      19216,
-		Component: 1,
-	}
-	hostLocal, err := NewCandidateHost(hostConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct local host candidate: %s", err)
-	}
-
-	srflxConfig := &CandidateServerReflexiveConfig{
-		Network:   "udp",
-		Address:   "192.168.1.1",
-		Port:      19217,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43212,
-	}
-	srflxLocal, err := NewCandidateServerReflexive(srflxConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct local srflx candidate: %s", err)
-	}
-
-	a.localCandidates[NetworkTypeUDP4] = []Candidate{hostLocal, srflxLocal}
-
-	localStats := a.GetLocalCandidatesStats()
-	if len(localStats) != 2 {
-		t.Fatalf("expected 2 local candidates stats, got %d instead", len(localStats))
-	}
-
-	var hostLocalStat, srflxLocalStat CandidateStats
-	for _, stats := range localStats {
-		var candidate Candidate
-		switch stats.ID {
-		case hostLocal.ID():
-			hostLocalStat = stats
-			candidate = hostLocal
-		case srflxLocal.ID():
-			srflxLocalStat = stats
-			candidate = srflxLocal
-		default:
-			t.Fatal("invalid local candidate ID")
-		}
-
-		if stats.CandidateType != candidate.Type() {
-			t.Fatal("invalid stats CandidateType")
-		}
-
-		if stats.Priority != candidate.Priority() {
-			t.Fatal("invalid stats CandidateType")
-		}
-
-		if stats.IP != candidate.Address() {
-			t.Fatal("invalid stats IP")
-		}
-	}
-
-	if hostLocalStat.ID != hostLocal.ID() {
-		t.Fatal("missing host local stat")
-	}
-
-	if srflxLocalStat.ID != srflxLocal.ID() {
-		t.Fatal("missing srflx local stat")
-	}
-
-	assert.NoError(t, a.Close())
-}
-
-func TestRemoteCandidateStats(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	// Avoid deadlocks?
-	defer test.TimeOut(1 * time.Second).Stop()
-
-	a, err := NewAgent(&AgentConfig{})
-	if err != nil {
-		t.Fatalf("Failed to create agent: %s", err)
-	}
-
-	relayConfig := &CandidateRelayConfig{
-		Network:   "udp",
-		Address:   "1.2.3.4",
-		Port:      12340,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43210,
-	}
-	relayRemote, err := NewCandidateRelay(relayConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote relay candidate: %s", err)
-	}
-
-	srflxConfig := &CandidateServerReflexiveConfig{
-		Network:   "udp",
-		Address:   "10.10.10.2",
-		Port:      19218,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43212,
-	}
-	srflxRemote, err := NewCandidateServerReflexive(srflxConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote srflx candidate: %s", err)
-	}
-
-	prflxConfig := &CandidatePeerReflexiveConfig{
-		Network:   "udp",
-		Address:   "10.10.10.2",
-		Port:      19217,
-		Component: 1,
-		RelAddr:   "4.3.2.1",
-		RelPort:   43211,
-	}
-	prflxRemote, err := NewCandidatePeerReflexive(prflxConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote prflx candidate: %s", err)
-	}
-
-	hostConfig := &CandidateHostConfig{
-		Network:   "udp",
-		Address:   "1.2.3.5",
-		Port:      12350,
-		Component: 1,
-	}
-	hostRemote, err := NewCandidateHost(hostConfig)
-	if err != nil {
-		t.Fatalf("Failed to construct remote host candidate: %s", err)
-	}
-
-	a.remoteCandidates[NetworkTypeUDP4] = []Candidate{relayRemote, srflxRemote, prflxRemote, hostRemote}
-
-	remoteStats := a.GetRemoteCandidatesStats()
-	if len(remoteStats) != 4 {
-		t.Fatalf("expected 4 remote candidates stats, got %d instead", len(remoteStats))
-	}
-	var relayRemoteStat, srflxRemoteStat, prflxRemoteStat, hostRemoteStat CandidateStats
-	for _, stats := range remoteStats {
-		var candidate Candidate
-		switch stats.ID {
-		case relayRemote.ID():
-			relayRemoteStat = stats
-			candidate = relayRemote
-		case srflxRemote.ID():
-			srflxRemoteStat = stats
-			candidate = srflxRemote
-		case prflxRemote.ID():
-			prflxRemoteStat = stats
-			candidate = prflxRemote
-		case hostRemote.ID():
-			hostRemoteStat = stats
-			candidate = hostRemote
-		default:
-			t.Fatal("invalid remote candidate ID")
-		}
-
-		if stats.CandidateType != candidate.Type() {
-			t.Fatal("invalid stats CandidateType")
-		}
-
-		if stats.Priority != candidate.Priority() {
-			t.Fatal("invalid stats CandidateType")
-		}
-
-		if stats.IP != candidate.Address() {
-			t.Fatal("invalid stats IP")
-		}
-	}
-
-	if relayRemoteStat.ID != relayRemote.ID() {
-		t.Fatal("missing relay remote stat")
-	}
-
-	if srflxRemoteStat.ID != srflxRemote.ID() {
-		t.Fatal("missing srflx remote stat")
-	}
-
-	if prflxRemoteStat.ID != prflxRemote.ID() {
-		t.Fatal("missing prflx remote stat")
-	}
-
-	if hostRemoteStat.ID != hostRemote.ID() {
-		t.Fatal("missing host remote stat")
-	}
-
-	assert.NoError(t, a.Close())
 }
 
 func TestInitExtIPMapping(t *testing.T) {
@@ -1316,415 +853,4 @@ func TestGetRemoteCredentials(t *testing.T) {
 	assert.Equal(t, actualPwd, a.remotePwd)
 
 	assert.NoError(t, a.Close())
-}
-
-func TestCloseInConnectionStateCallback(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 5)
-	defer lim.Stop()
-
-	disconnectedDuration := time.Second
-	failedDuration := time.Second
-	KeepaliveInterval := time.Duration(0)
-	CheckInterval := 500 * time.Millisecond
-
-	cfg := &AgentConfig{
-		Urls:                []*stun.URI{},
-		NetworkTypes:        supportedNetworkTypes(),
-		DisconnectedTimeout: &disconnectedDuration,
-		FailedTimeout:       &failedDuration,
-		KeepaliveInterval:   &KeepaliveInterval,
-		CheckInterval:       &CheckInterval,
-	}
-
-	aAgent, err := NewAgent(cfg)
-	if err != nil {
-		t.Error(err)
-	}
-
-	bAgent, err := NewAgent(cfg)
-	if err != nil {
-		t.Error(err)
-	}
-
-	isClosed := make(chan interface{})
-	isConnected := make(chan interface{})
-	err = aAgent.OnConnectionStateChange(func(c ConnectionState) {
-		switch c {
-		case ConnectionStateConnected:
-			<-isConnected
-			assert.NoError(t, aAgent.Close())
-		case ConnectionStateClosed:
-			close(isClosed)
-		default:
-		}
-	})
-	if err != nil {
-		t.Error(err)
-	}
-
-	connect(aAgent, bAgent)
-	close(isConnected)
-
-	<-isClosed
-	assert.NoError(t, bAgent.Close())
-}
-
-func TestRunTaskInConnectionStateCallback(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 5)
-	defer lim.Stop()
-
-	oneSecond := time.Second
-	KeepaliveInterval := time.Duration(0)
-	CheckInterval := 50 * time.Millisecond
-
-	cfg := &AgentConfig{
-		Urls:                []*stun.URI{},
-		NetworkTypes:        supportedNetworkTypes(),
-		DisconnectedTimeout: &oneSecond,
-		FailedTimeout:       &oneSecond,
-		KeepaliveInterval:   &KeepaliveInterval,
-		CheckInterval:       &CheckInterval,
-	}
-
-	aAgent, err := NewAgent(cfg)
-	check(err)
-	bAgent, err := NewAgent(cfg)
-	check(err)
-
-	isComplete := make(chan interface{})
-	err = aAgent.OnConnectionStateChange(func(c ConnectionState) {
-		if c == ConnectionStateConnected {
-			_, _, errCred := aAgent.GetLocalUserCredentials()
-			assert.NoError(t, errCred)
-			assert.NoError(t, aAgent.Restart("", ""))
-			close(isComplete)
-		}
-	})
-	if err != nil {
-		t.Error(err)
-	}
-
-	connect(aAgent, bAgent)
-
-	<-isComplete
-	assert.NoError(t, aAgent.Close())
-	assert.NoError(t, bAgent.Close())
-}
-
-func TestRunTaskInSelectedCandidatePairChangeCallback(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 5)
-	defer lim.Stop()
-
-	oneSecond := time.Second
-	KeepaliveInterval := time.Duration(0)
-	CheckInterval := 50 * time.Millisecond
-
-	cfg := &AgentConfig{
-		Urls:                []*stun.URI{},
-		NetworkTypes:        supportedNetworkTypes(),
-		DisconnectedTimeout: &oneSecond,
-		FailedTimeout:       &oneSecond,
-		KeepaliveInterval:   &KeepaliveInterval,
-		CheckInterval:       &CheckInterval,
-	}
-
-	aAgent, err := NewAgent(cfg)
-	check(err)
-	bAgent, err := NewAgent(cfg)
-	check(err)
-
-	isComplete := make(chan interface{})
-	isTested := make(chan interface{})
-	if err = aAgent.OnSelectedCandidatePairChange(func(Candidate, Candidate) {
-		go func() {
-			_, _, errCred := aAgent.GetLocalUserCredentials()
-			assert.NoError(t, errCred)
-			close(isTested)
-		}()
-	}); err != nil {
-		t.Error(err)
-	}
-	if err = aAgent.OnConnectionStateChange(func(c ConnectionState) {
-		if c == ConnectionStateConnected {
-			close(isComplete)
-		}
-	}); err != nil {
-		t.Error(err)
-	}
-
-	connect(aAgent, bAgent)
-
-	<-isComplete
-	<-isTested
-	assert.NoError(t, aAgent.Close())
-	assert.NoError(t, bAgent.Close())
-}
-
-// Assert that a Lite agent goes to disconnected and failed
-func TestLiteLifecycle(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 30)
-	defer lim.Stop()
-
-	aNotifier, aConnected := onConnected()
-
-	aAgent, err := NewAgent(&AgentConfig{
-		NetworkTypes:     supportedNetworkTypes(),
-		MulticastDNSMode: MulticastDNSModeDisabled,
-	})
-	require.NoError(t, err)
-	require.NoError(t, aAgent.OnConnectionStateChange(aNotifier))
-
-	disconnectedDuration := time.Second
-	failedDuration := time.Second
-	KeepaliveInterval := time.Duration(0)
-	CheckInterval := 500 * time.Millisecond
-	bAgent, err := NewAgent(&AgentConfig{
-		Lite:                true,
-		CandidateTypes:      []CandidateType{CandidateTypeHost},
-		NetworkTypes:        supportedNetworkTypes(),
-		MulticastDNSMode:    MulticastDNSModeDisabled,
-		DisconnectedTimeout: &disconnectedDuration,
-		FailedTimeout:       &failedDuration,
-		KeepaliveInterval:   &KeepaliveInterval,
-		CheckInterval:       &CheckInterval,
-	})
-	require.NoError(t, err)
-
-	bConnected := make(chan interface{})
-	bDisconnected := make(chan interface{})
-	bFailed := make(chan interface{})
-
-	require.NoError(t, bAgent.OnConnectionStateChange(func(c ConnectionState) {
-		fmt.Println(c)
-		switch c {
-		case ConnectionStateConnected:
-			close(bConnected)
-		case ConnectionStateDisconnected:
-			close(bDisconnected)
-		case ConnectionStateFailed:
-			close(bFailed)
-		default:
-		}
-	}))
-
-	connectWithVNet(bAgent, aAgent)
-
-	<-aConnected
-	<-bConnected
-	assert.NoError(t, aAgent.Close())
-
-	<-bDisconnected
-	<-bFailed
-	assert.NoError(t, bAgent.Close())
-}
-
-func TestNilCandidate(t *testing.T) {
-	a, err := NewAgent(&AgentConfig{})
-	assert.NoError(t, err)
-
-	assert.NoError(t, a.AddRemoteCandidate(nil))
-	assert.NoError(t, a.Close())
-}
-
-func TestNilCandidatePair(t *testing.T) {
-	a, err := NewAgent(&AgentConfig{})
-	assert.NoError(t, err)
-
-	a.setSelectedPair(nil)
-	assert.NoError(t, a.Close())
-}
-
-func TestGetSelectedCandidatePair(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 30)
-	defer lim.Stop()
-
-	wan, err := vnet.NewRouter(&vnet.RouterConfig{
-		CIDR:          "0.0.0.0/0",
-		LoggerFactory: logging.NewDefaultLoggerFactory(),
-	})
-	assert.NoError(t, err)
-
-	net, err := vnet.NewNet(&vnet.NetConfig{
-		StaticIPs: []string{"192.168.0.1"},
-	})
-	assert.NoError(t, err)
-	assert.NoError(t, wan.AddNet(net))
-
-	assert.NoError(t, wan.Start())
-
-	cfg := &AgentConfig{
-		NetworkTypes: supportedNetworkTypes(),
-		Net:          net,
-	}
-
-	aAgent, err := NewAgent(cfg)
-	assert.NoError(t, err)
-
-	bAgent, err := NewAgent(cfg)
-	assert.NoError(t, err)
-
-	aAgentPair, err := aAgent.GetSelectedCandidatePair()
-	assert.NoError(t, err)
-	assert.Nil(t, aAgentPair)
-
-	bAgentPair, err := bAgent.GetSelectedCandidatePair()
-	assert.NoError(t, err)
-	assert.Nil(t, bAgentPair)
-
-	connect(aAgent, bAgent)
-
-	aAgentPair, err = aAgent.GetSelectedCandidatePair()
-	assert.NoError(t, err)
-	assert.NotNil(t, aAgentPair)
-
-	bAgentPair, err = bAgent.GetSelectedCandidatePair()
-	assert.NoError(t, err)
-	assert.NotNil(t, bAgentPair)
-
-	assert.True(t, bAgentPair.Local.Equal(aAgentPair.Remote))
-	assert.True(t, bAgentPair.Remote.Equal(aAgentPair.Local))
-
-	assert.NoError(t, wan.Stop())
-	assert.NoError(t, aAgent.Close())
-	assert.NoError(t, bAgent.Close())
-}
-
-func TestAcceptAggressiveNomination(t *testing.T) {
-	report := test.CheckRoutines(t)
-	defer report()
-
-	lim := test.TimeOut(time.Second * 30)
-	defer lim.Stop()
-
-	// Create a network with two interfaces
-	wan, err := vnet.NewRouter(&vnet.RouterConfig{
-		CIDR:          "0.0.0.0/0",
-		LoggerFactory: logging.NewDefaultLoggerFactory(),
-	})
-	assert.NoError(t, err)
-
-	net0, err := vnet.NewNet(&vnet.NetConfig{
-		StaticIPs: []string{"192.168.0.1"},
-	})
-	assert.NoError(t, err)
-	assert.NoError(t, wan.AddNet(net0))
-
-	net1, err := vnet.NewNet(&vnet.NetConfig{
-		StaticIPs: []string{"192.168.0.2", "192.168.0.3", "192.168.0.4"},
-	})
-	assert.NoError(t, err)
-	assert.NoError(t, wan.AddNet(net1))
-
-	assert.NoError(t, wan.Start())
-
-	aNotifier, aConnected := onConnected()
-	bNotifier, bConnected := onConnected()
-
-	KeepaliveInterval := time.Hour
-	cfg0 := &AgentConfig{
-		NetworkTypes:     []NetworkType{NetworkTypeUDP4, NetworkTypeUDP6},
-		MulticastDNSMode: MulticastDNSModeDisabled,
-		Net:              net0,
-
-		KeepaliveInterval:          &KeepaliveInterval,
-		CheckInterval:              &KeepaliveInterval,
-		AcceptAggressiveNomination: true,
-	}
-
-	var aAgent, bAgent *Agent
-	aAgent, err = NewAgent(cfg0)
-	require.NoError(t, err)
-	require.NoError(t, aAgent.OnConnectionStateChange(aNotifier))
-
-	cfg1 := &AgentConfig{
-		NetworkTypes:      []NetworkType{NetworkTypeUDP4, NetworkTypeUDP6},
-		MulticastDNSMode:  MulticastDNSModeDisabled,
-		Net:               net1,
-		KeepaliveInterval: &KeepaliveInterval,
-		CheckInterval:     &KeepaliveInterval,
-	}
-
-	bAgent, err = NewAgent(cfg1)
-	require.NoError(t, err)
-	require.NoError(t, bAgent.OnConnectionStateChange(bNotifier))
-
-	aConn, bConn := connect(aAgent, bAgent)
-
-	// Ensure pair selected
-	// Note: this assumes ConnectionStateConnected is thrown after selecting the final pair
-	<-aConnected
-	<-bConnected
-
-	// Send new USE-CANDIDATE message with higher priority to update the selected pair
-	buildMsg := func(class stun.MessageClass, username, key string, priority uint32) *stun.Message {
-		msg, err1 := stun.Build(stun.NewType(stun.MethodBinding, class), stun.TransactionID,
-			stun.NewUsername(username),
-			stun.NewShortTermIntegrity(key),
-			UseCandidate(),
-			PriorityAttr(priority),
-			stun.Fingerprint,
-		)
-		if err1 != nil {
-			t.Fatal(err1)
-		}
-
-		return msg
-	}
-
-	selectedCh := make(chan Candidate, 1)
-	var expectNewSelectedCandidate Candidate
-	err = aAgent.OnSelectedCandidatePairChange(func(_, remote Candidate) {
-		selectedCh <- remote
-	})
-	require.NoError(t, err)
-	var bcandidates []Candidate
-	bcandidates, err = bAgent.GetLocalCandidates()
-	require.NoError(t, err)
-
-	for _, c := range bcandidates {
-		if c != bAgent.getSelectedPair().Local {
-			if expectNewSelectedCandidate == nil {
-			incr_priority:
-				for _, candidates := range aAgent.remoteCandidates {
-					for _, candidate := range candidates {
-						if candidate.Equal(c) {
-							candidate.(*CandidateHost).priorityOverride += 1000 //nolint:forcetypeassert
-							break incr_priority
-						}
-					}
-				}
-				expectNewSelectedCandidate = c
-			}
-			_, err = c.writeTo(buildMsg(stun.ClassRequest, aAgent.localUfrag+":"+aAgent.remoteUfrag, aAgent.localPwd, c.Priority()).Raw, bAgent.getSelectedPair().Remote)
-			require.NoError(t, err)
-		}
-	}
-
-	time.Sleep(1 * time.Second)
-	select {
-	case selected := <-selectedCh:
-		assert.True(t, selected.Equal(expectNewSelectedCandidate))
-	default:
-		t.Fatal("No selected candidate pair")
-	}
-
-	assert.NoError(t, wan.Stop())
-	if !closePipe(t, aConn, bConn) {
-		return
-	}
 }
